+++ src/modules/kapur/v1/farmer/farmer.model.ts
@@ -19,7 +19,6 @@ const farmerSchema = new Schema<IFarmer>(
     name: {
       type: String,
       required: true,
-      index: true,
       trim: true,
     },